cli parser improved
Used the clap crater to create a cli parser that is more powerful and instructive than the previous parser that was written manually.
The parser can now parse commands and command arguments more easily (and it is easy to add commands and parameters). This allowed to make sure that Add can get all the parameters required for a todo item.